feat: 회원 가입 기능 구현 완료 1. 회원 가입 및 이메일 중복 검사 TDD 검증 완료 2. ResponseDTO body에 null값 보내올 경우 isSuccess필드만 넘겨주도록 수정

+++ src/authservice/src/main/java/com/eventty/authservice/domain/entity/AuthorityEntity.java
@@ -1,10 +1,7 @@
 package com.eventty.authservice.domain.entity;
 
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 @Getter
 @Entity
@@ -16,7 +13,7 @@ public class AuthorityEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String name;    // 권한 이름(역할)
